improvement(blocks): extract model config subBlocks into shared utils
